apps/scrapers-python/requirements.txt
# Web Scraping
scrapy>=2.11.0            # Legacy — being replaced by Scrapling in scrapers/
scrapling[fetchers]>=0.4  # Adaptive scraping (Phase 3 migration target)
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.45.0

# PDF Processing
pdfplumber>=0.10.0
pymupdf>=1.24.0  # fitz

# Data Processing
pandas>=2.2.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Redis/Queue
redis>=5.0.0
rq>=1.16.0

# AI/Embeddings (Together.ai - uses requests for HTTP calls)
# No extra SDK needed - uses requests which is already listed below

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0
aiohttp>=3.9.0
tenacity>=8.4.0  # Retry logic
